+++ CHANGELOG.markdown
@@ -4,6 +4,7 @@
   `mkTransformationMat`, `basisFor`, `scaled` by using implementations
   that inline well for functions that were previously reference
   implementations.
+* Add type and pattern synonyms for P0, P1, P2, P3 and P4.
 
 1.16.2
 ----